requirements.txt
selenium
webdriver-manager
beautifulsoup4
requests
python-dotenv